+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2019-04-19
+
+### Fixed
+
+- Packaging script typo that prevended NPM interoperability (thanks to [@a6kme](https://github.com/a6kme)).
+
+### Changed
+
+- Android `getBitmapFromUriOrDrawable` has been refactored for better maintainability.
+- Android `convertDpToPixel` has been replaced with Facebook's `PixelUtil.toPixelFromDIP` since it's more tested and reliable.
+
 ## [1.4.0] - 2019-03-01
 
 ### Added
@@ -115,6 +126,7 @@ The following functions have been deprecated. Please see
 
 - `openInApp`: open a valid http(s) URL with an in-app browser.
 
+[1.4.1]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.2.7...v1.3.0
 [1.2.7]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.2.4...v1.2.7